+++ src/decorators/index.ts
@@ -1,3 +1,2 @@
-export * from './mutation-with-zod'
-export * from './query-with-zod'
-export * from './subscription-with-zod'
+export * from './query-types'
+export * from './input-type'