+++ src/decorators/query-types/query-with-zod.ts
@@ -2,8 +2,8 @@ import * as zod from 'zod'
 
 import { Query, QueryOptions as QO } from '@nestjs/graphql'
 
-import { IModelFromZodOptionsWithMapper } from '../model-from-zod'
-import { MethodWithZod } from './common'
+import { IModelFromZodOptionsWithMapper } from '../../model-from-zod'
+import { MethodWithZod } from '../common'
 
 export interface QueryOptions<T extends object> extends QO {
   /**
@@ -51,10 +51,10 @@ export function QueryWithZod<T extends zod.AnyZodObject>(input: T, name: string)
  * @export
  * @template T The type of the zod object input.
  * @param {T} input The zod input object.
- * @param {QueryOptions<zod.infer<T>>} options The options for query.
+ * @param {QueryOptions<T>} options The options for query.
  * @return {MethodDecorator} A {@link MethodDecorator}.
  */
-export function QueryWithZod<T extends zod.AnyZodObject>(input: T, options: QueryOptions<zod.infer<T>>): MethodDecorator
-export function QueryWithZod<T extends zod.AnyZodObject>(input: T, nameOrOptions?: string | QueryOptions<zod.infer<T>>) {
-  return MethodWithZod(input, nameOrOptions!, Query)
+export function QueryWithZod<T extends zod.AnyZodObject>(input: T, options: QueryOptions<T>): MethodDecorator
+export function QueryWithZod<T extends zod.AnyZodObject>(input: T, nameOrOptions?: string | QueryOptions<T>) {
+  return MethodWithZod(input, nameOrOptions, Query)
 }