+++ src/index.ts
@@ -1,3 +1,8 @@
 export * from './decorators'
 export { getZodObject } from './helpers'
-export { IModelFromZodOptions, modelFromZod } from './model-from-zod'
+
+export {
+  IModelFromZodOptions,
+  modelFromZod,
+  modelFromZodBase,
+} from './model-from-zod'